requirements.txt
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.6.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-mail==1.4.1
greenlet==3.1.0
h11==0.14.0
hiredis==3.0.0
idna==3.10
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
redis==5.1.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.30.6
